modified:   adscompstat/tasks.py  	new file:   adscompstat/tasks_old.py  	modified:   config.py
