+++ config.py
@@ -1,13 +1,13 @@
-HARVEST_BASE_DIR = '/data/Crossref/'
+HARVEST_BASE_DIR = '/app/data/Crossref/'
 
 HARVEST_LOG_DIR = HARVEST_BASE_DIR + '/UpdateAgent/'
 
-CLASSIC_DOI_FILE = '/data/all.links'
-CLASSIC_ALTBIBS = '/data/bibcodes.list.alt'
-CLASSIC_DELBIBS = '/data/bibcodes.list.del'
-CLASSIC_ALLBIBS = '/data/bibcodes.list.all'
-CLASSIC_CANONICAL = '/data/bibcodes.list.can'
+CLASSIC_DOI_FILE = '/app/data/all.links'
+CLASSIC_ALTBIBS = '/app/data/bibcodes.list.alt'
+CLASSIC_DELBIBS = '/app/data/bibcodes.list.del'
+CLASSIC_ALLBIBS = '/app/data/bibcodes.list.all'
+CLASSIC_CANONICAL = '/app/data/bibcodes.list.can'
 
-JOURNALSDB_ISSN_BIBSTEM = '/data/issn_identifiers'
+JOURNALSDB_ISSN_BIBSTEM = '/app/data/issn_identifiers'
 
-COMPLETENESS_EXPORT_FILE = '/data/completeness_export.json'
+COMPLETENESS_EXPORT_FILE = '/app/data/completeness_export.json'
